feat: routing fixes, memory fixes
chore: clippy

chore: data clippy

chore: clippy memory

chore: cargo format

chore: editions

chore: add derive to bytemuck features

chore: fmt

+++ crates/memory/src/game_engine/unity/mod.rs
@@ -55,7 +55,7 @@ impl Module {
         };
 
         let assemblies = process
-            .read_pointer::<u64>(assemblies_pointer as u64)
+            .read_pointer::<u64>(assemblies_pointer)
             .ok()
             .filter(|val| *val != 0)?;
